LayoutTests/media/modern-media-controls/media-controls/media-controls-placard.html

<script src="../../../resources/js-test-pre.js"></script>
<script src="../resources/media-controls-loader.js" type="text/javascript"></script>
<body>
<script type="text/javascript">

description("Testing the <code>MediaControls</code> handling of placards.");

const mediaControls = new MediaControls({ layoutTraits: LayoutTraits.macOS });

debug("By default, no placard is shown.");
shouldBeFalse("mediaControls.showsPlacard");
shouldBe("mediaControls.children", "[mediaControls.controlsBar]");

debug("");
debug("Showing the Airplay placard");
mediaControls.showPlacard(mediaControls.airplayPlacard);
shouldBeTrue("mediaControls.showsPlacard");
shouldBe("mediaControls.children", "[mediaControls.airplayPlacard, mediaControls.controlsBar]");

debug("");
debug("Showing the Invalid placard, hides the controls bar");
mediaControls.showPlacard(mediaControls.invalidPlacard);
shouldBeTrue("mediaControls.showsPlacard");
shouldBe("mediaControls.children", "[mediaControls.invalidPlacard]");

debug("");
debug("Showing the PiP placard, hides the controls bar");
mediaControls.showPlacard(mediaControls.pipPlacard);
shouldBeTrue("mediaControls.showsPlacard");
shouldBe("mediaControls.children", "[mediaControls.pipPlacard]");

debug("");
debug("Hide placard");
mediaControls.hidePlacard();
shouldBeFalse("mediaControls.showsPlacard");
shouldBe("mediaControls.children", "[mediaControls.controlsBar]");

debug("");

</script>
<script src="../../../resources/js-test-post.js"></script>
</body>
