LayoutTests/media/modern-media-controls/media-controls/media-controls-placard-expected.txt

Testing the MediaControls handling of placards.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


By default, no placard is shown.
PASS mediaControls.showsPlacard is false
PASS mediaControls.children is [mediaControls.controlsBar]

Showing the Airplay placard
PASS mediaControls.showsPlacard is true
PASS mediaControls.children is [mediaControls.airplayPlacard, mediaControls.controlsBar]

Showing the Invalid placard, hides the controls bar
PASS mediaControls.showsPlacard is true
PASS mediaControls.children is [mediaControls.invalidPlacard]

Showing the PiP placard, hides the controls bar
PASS mediaControls.showsPlacard is true
PASS mediaControls.children is [mediaControls.pipPlacard]

Hide placard
PASS mediaControls.showsPlacard is false
PASS mediaControls.children is [mediaControls.controlsBar]

PASS successfullyParsed is true

TEST COMPLETE
